move components folder from app to src
