+++ app/mercadona/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import ProductList from '../components/ProductList/ProductList'
+import ProductList from '../../src/components/ProductList/ProductList'
 import { SuperMarkets } from '@/src/types/superMarkets'
 
 export default function Mercadona() {
