+++ src/components/cool-card/cool-card.tsx
